backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
requests
youtube-transcript-api
groq
python-dotenv
yt-dlp
